+++ main-app/client/src/components/Intake/ContactInfoFormGroup.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Field } from 'formik'
-import '../style/FormGroup.scss'
-import BasicField from '../../shared/BasicField'
+import '../Form/IntakeForm/style/FormGroup.scss'
+import { BasicField } from '../Form/shared'
 
 const starStyle = {
   color: '#f44336',
